+++ packages/exchange/src/server/index.ts
@@ -1,8 +1,0 @@
-// deprecated methods
-export { SockJsFakeHost, WsFakeHost } from '../deprecated'
-export type { FakeHost, ProtocolHandler, ServiceDefinition } from '../deprecated'
-
-// new methods
-export * from '../ws/WsHost'
-export { type Host } from '../ws/Host'
-export * from '../logger'