+++ packages/exchange/tsup.config.ts
@@ -2,11 +2,8 @@ import { defineConfig } from 'tsup'
 
 export default defineConfig({
     format: ['esm', 'cjs'],
-    dts: {
-        resolve: true,
-    },
-    entry: ['src/index.ts', 'src/browser', 'src/server'],
-    splitting: true,
-    sourcemap: true,
+    dts: 'src/index.ts',
+    entry: ['src/index.ts', 'src/browser.ts'],
+    splitting: false,
     clean: true,
 })