+++ packages/rest/fake-rest/src/host/HttpRestServiceStub.ts
@@ -1,0 +1,15 @@
+import { HttpRest } from './types'
+import { logger } from '../logger'
+
+export class HttpRestServiceStub implements HttpRest {
+    public readonly server = null
+    public readonly url = Promise.resolve(new URL('http://localhost'))
+
+    constructor() {
+        logger('HttpRestService does not run in browser environment')
+    }
+
+    dispose() {
+        return
+    }
+}