+++ packages/rest/fake-rest/tsup.config.ts
@@ -2,15 +2,8 @@ import { defineConfig } from 'tsup'
 
 export default defineConfig({
     format: ['esm', 'cjs'],
-    dts: {
-        resolve: true,
-    },
-    entry: {
-        index: 'src/index.ts',
-        server: 'src/HttpRestService.ts',
-        browser: 'src/HijackedRestService.ts',
-    },
-    splitting: true,
-    sourcemap: true,
+    dts: 'src/index.ts',
+    entry: ['src/index.ts', 'src/browser.ts'],
+    splitting: false,
     clean: true,
 })