+++ packages/rest/test-fake-svc/src/testSetup.ts
@@ -1,6 +1,5 @@
-import { HttpRestService, createRouter } from '@fakehost/fake-rest/server'
+import { HttpRestService, createRouter } from '@fakehost/fake-rest'
 import { Configuration } from '@fakehost/rest-generated-client-api'
-import { URL } from 'url'
 
 export type TestTarget = 'FAKE' | 'REMOTE'