+++ packages/signalr/fake-signalr/src/index.ts
@@ -1,2 +1,6 @@
-export { FakeSignalrHub } from './FakeSignalrHub'
 export type { ConnectionId } from '@fakehost/exchange'
+export { FakeSignalrHub } from './FakeSignalrHub'
+export * from './create/browserSignalr'
+export { enableLogger } from './create/logger'
+export * from './create/serverSignalr'
+export type { ServerOptions } from './types'