+++ packages/signalr/fake-signalr/src/types.ts
@@ -1,5 +1,12 @@
-import Url from 'url'
 import { FakeSignalrHub } from './FakeSignalrHub'
+
+export type ServerOptions<T extends Record<string, unknown>> = {
+    url: URL
+    name?: string
+    silent?: boolean
+    debug?: boolean
+    hubs: T
+}
 
 export type Signalr<T> = T extends FakeSignalrHub<infer H, infer R, infer S>
     ? FakeSignalrHub<H, R, S>
@@ -19,4 +26,4 @@ export const isFakeSignalrHub = (hub: any): hub is FakeSignalrHub<any, any, any>
     )
 }
 
-export const URL = globalThis.URL || Url.URL
+export const URL = globalThis.URL