+++ packages/signalr/fake-signalr/CHANGELOG.md
@@ -3,6 +3,98 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.14](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.13...@fakehost/signalr@1.0.0-beta.14) (2024-11-20)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.13](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.12...@fakehost/signalr@1.0.0-beta.13) (2024-11-18)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.12](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.11...@fakehost/signalr@1.0.0-beta.12) (2024-11-18)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.11](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.10...@fakehost/signalr@1.0.0-beta.11) (2024-11-18)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.10](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.9...@fakehost/signalr@1.0.0-beta.10) (2024-11-15)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.9](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.8...@fakehost/signalr@1.0.0-beta.9) (2024-11-15)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.8](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.7...@fakehost/signalr@1.0.0-beta.8) (2024-11-15)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.7](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.6...@fakehost/signalr@1.0.0-beta.7) (2024-11-15)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.6](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.5...@fakehost/signalr@1.0.0-beta.6) (2024-11-15)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.5](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.4...@fakehost/signalr@1.0.0-beta.5) (2024-11-15)
+
+**Note:** Version bump only for package @fakehost/signalr
+
+
+
+
+
+# [1.0.0-beta.4](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.3...@fakehost/signalr@1.0.0-beta.4) (2024-10-24)
+
+
+### Bug Fixes
+
+* browser/node imports ([c140330](https://github.com/ilikejames/fakehost/commit/c1403306e99ab1e4bf4884c42cfde9baa890ac4e))
+* embedded types ([d9d456f](https://github.com/ilikejames/fakehost/commit/d9d456f3bb2cf4d41979a6ede4786b41c6108649))
+
+
+
+
+
 # [1.0.0-beta.3](https://github.com/ilikejames/fakehost/compare/@fakehost/signalr@1.0.0-beta.2...@fakehost/signalr@1.0.0-beta.3) (2023-10-04)
 
 **Note:** Version bump only for package @fakehost/signalr